requirements.txt
Django==1.6.8
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.2
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.19
celery==3.1.17
django-celery==3.1.16
django-celery-with-redis==3.0
django-debug-toolbar==1.2.2
django-devserver==0.8.0
django-extensions==1.4.8
django-faker==0.2
django-reversion==1.8.5
django-versioning==0.7.5.0
django-autofixture==0.9.1
django-mptt==0.6.1
django-cte-trees==1.0b1
djorm-ext-core==0.7
djorm-ext-expressions==0.6
fake-factory==0.4.2
flower==0.7.3
kombu==3.0.24
psycopg2==2.5.4
pytz==2014.10
redis==2.10.3
six==1.8.0
sqlparse==0.1.14
tornado==4.0.2
wsgiref==0.1.2
django-hstore==1.3.5
django-jsonfield==0.9.13

# may be remove
django-json-dbindex==1.1.0
